Rename file for validateNodeEngineVersion
Fix the descriptions in the specs
